fixing warnings, refs #3, #2495
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@21206 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/gui/GUIApplicationWindow.cpp
@@ -1520,10 +1520,10 @@ GUIApplicationWindow::closeAllWindows() {
     }
     // remove trackers and other external windows
     int i;
-    for (i = 0; i < mySubWindows.size(); ++i) {
+    for (i = 0; i < (int)mySubWindows.size(); ++i) {
         mySubWindows[i]->destroy();
     }
-    for (i = 0; i < myTrackerWindows.size(); ++i) {
+    for (i = 0; i < (int)myTrackerWindows.size(); ++i) {
         myTrackerWindows[i]->destroy();
     }
     // delete the simulation
