+++ sumo/src/gui/GUIViewTraffic.cpp
@@ -285,7 +285,7 @@ GUIViewTraffic::onGamingClick(Position pos) {
         const std::vector<MSTrafficLightLogic*> logics = vars.getAllLogics();
         if (logics.size() > 1) {
             MSSimpleTrafficLightLogic* l = (MSSimpleTrafficLightLogic*) logics[0];
-            for (int i = 0; i < logics.size() - 1; ++i) {
+            for (int i = 0; i < (int)logics.size() - 1; ++i) {
                 if (minTll->getProgramID() == logics[i]->getProgramID()) {
                     l = (MSSimpleTrafficLightLogic*) logics[i + 1];
                     tlsControl.switchTo(minTll->getID(), l->getProgramID());
